requirements.txt
# Python Backend Dependencies
# These packages are only needed if you want to run the data collection scripts

tuya-connector-python==0.1.2
requests>=2.31.0
python-dateutil>=2.8.2
fastapi==0.115.0
uvicorn==0.30.6
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
